4-Moves helper classes and models to server directory
